svg/painting/parsing/stroke-linejoin-invalid.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml" width="800" height="800"><title>SVG Painting: parsing stroke-linejoin with invalid values</title><metadata><h:link href="https://svgwg.org/svg2-draft/painting.html#StrokeLinejoinProperty" rel="help"/><h:meta content="stroke-linejoin supports only the grammar 'miter | miter-clip | round | bevel | arcs'" name="assert"/></metadata><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><h:script src="/css/support/parsing-testcommon.js"/><script><![CDATA[

test_invalid_value("stroke-linejoin", "auto");
test_invalid_value("stroke-linejoin", "round bevel");

  ]]></script></svg>
